# PERF: Series.str.get_dummies for ArrowDtype(pa.string())