- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

```
import pandas as pd
import pyarrow as pa

data = ["a|b|c", "a|b", "a|c", "b|c", "a", "b", "c"] * 1000
ser = pd.Series(data, dtype=pd.ArrowDtype(pa.string()))

%timeit ser.str.get_dummies() 

# 549 ms ± 43.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)     -> main
# 8.67 ms ± 415 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  -> PR
```