@mroeschke 2023-06-14
Nice optimizations as usual. Thanks @lukemanley